requirements.txt
python-dotenv==1.0.1
pyTelegramBotAPI==4.12.0
yt-dlp==2023.7.6
ffmpeg-python==0.2.0
requests==2.31.0
